+++ CHANGES.md
@@ -10,6 +10,10 @@ Use list notation, and following prefixes:
 - Bugfix - when fixing any major bug
 - Docs - for any improvement to documentation
 
+### Changes:
+- Fix: No size availble text instead max size issue
+- Fix: Virtusize widget recommended size incorrect
+
 ### 2.12.25
 - Fix: Prevent send event userSawProduct when productValid = false or repeated product load